+++ src/export/ExportItem.cpp
@@ -1,4 +1,4 @@
-#include "export/exportItem.h"
+#include "export/ExportItem.h"
 MayaUSDExport::ExportItem::ExportItem(MDagPath dagPath)
 : dagPath(dagPath)
 {
